Second Codex round: demo masking, narrow reads, self-healing cache
- The Logistics tab POST now applies demo-mode overrides before anything is
  parsed or stored: the address is masked and the pin cleared, so demo
  instances never persist a real location.
- Other Attendees labels its rows with getListingNamesByIds instead of
  warming the whole listings cache.
- Address-cache rows from before coordinates existed now read as a miss, so
  the next search re-fetches the postcode WITH coordinates instead of
  serving locationless results until the row expires.
- Split the Logistics tab tests into two files (each within the size
  guideline) so their setup stays inside the suite's repeated-read budget.
