+++ src/features/admin/attendee-logistics-routes.ts
@@ -33,6 +33,7 @@ import type { TypedRouteHandler } from "#routes/router.ts";
 import { logActivity } from "#shared/db/activityLog.ts";
 import { updateAttendeePII } from "#shared/db/attendees.ts";
 import { setLogisticsAssignments } from "#shared/db/logistics.ts";
+import { applyDemoOverrides, LOGISTICS_DEMO_FIELDS } from "#shared/demo.ts";
 import type { FormParams } from "#shared/form-data.ts";
 import { parseCoordinatePair } from "#shared/validation/coordinates.ts";
 import { AttendeeLogisticsPanel } from "#templates/admin/attendee-logistics-tab.tsx";
@@ -99,6 +100,10 @@ const handleLogisticsSubmit = async (
   session: AuthSession,
   form: FormParams,
 ): Promise<Response> => {
+  // On demo instances the address is masked and the pin cleared before
+  // anything is parsed or stored, like the attendee form's own submitters.
+  applyDemoOverrides(form, LOGISTICS_DEMO_FIELDS);
+
   const entity = await loadAttendeeForEdit(attendeeId);
   if (!entity) return notFoundResponse();
